+++ src/main/java/com/debateseason_backend_v1/domain/profile/service/request/ProfileRegisterServiceRequest.java
@@ -1,0 +1,17 @@
+package com.debateseason_backend_v1.domain.profile.service.request;
+
+import com.debateseason_backend_v1.domain.profile.enums.AgeRangeType;
+import com.debateseason_backend_v1.domain.profile.enums.GenderType;
+
+import lombok.Builder;
+
+@Builder
+public record ProfileRegisterServiceRequest(
+	Long userId,
+	String profileColor,
+	String nickname,
+	Long communityId,
+	GenderType gender,
+	AgeRangeType ageRange
+) {
+}
