+++ src/main/java/com/debateseason_backend_v1/domain/profile/service/response/CommunityResponse.java
@@ -1,4 +1,4 @@
-package com.debateseason_backend_v1.domain.community.service.response;
+package com.debateseason_backend_v1.domain.profile.service.response;
 
 import com.debateseason_backend_v1.domain.repository.entity.Community;
